Aula: Implementando outras exceções

+++ src/main/java/com/pessoal/dscommerce/services/ProductService.java
@@ -1,15 +1,20 @@
 package com.pessoal.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.pessoal.dscommerce.dto.ProductDTO;
 import com.pessoal.dscommerce.entities.Product;
 import com.pessoal.dscommerce.repositories.ProductRepository;
+import com.pessoal.dscommerce.services.exceptions.DatabaseException;
 import com.pessoal.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -41,16 +46,27 @@ public class ProductService {
 
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		Product product = repository.getReferenceById(id);
-		copyDTOToEntity(dto, product);
-		product = repository.save(product);
-		return new ProductDTO(product);
+		try {
+			Product product = repository.getReferenceById(id);
+			copyDTOToEntity(dto, product);
+			product = repository.save(product);
+			return new ProductDTO(product);
 
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 	}
 
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+			repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial");
+		}
 	}
 
 	private void copyDTOToEntity(ProductDTO dto, Product entity) {
